Started splitting of client
