+++ dacs7/src/Dacs7/Dacs7.csproj
@@ -23,9 +23,9 @@
     <GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
     <GenerateAssemblyCopyrightAttribute>false</GenerateAssemblyCopyrightAttribute>
     <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-    <Version>1.4.1-beta1</Version>
-    <AssemblyVersion>1.4.1.0</AssemblyVersion>
-    <FileVersion>1.4.1.0</FileVersion>
+    <Version>2.0.0-beta1</Version>
+    <AssemblyVersion>2.0.0.0</AssemblyVersion>
+    <FileVersion>2.0.0.0</FileVersion>
     <RepositoryUrl>git://github.com/proemmer/dacs7</RepositoryUrl>
     <RepositoryType>git</RepositoryType>
   </PropertyGroup>
